w3c-reports/12092014/gtk/runned-tests/XMLHttpRequest/send-authentication-basic-cors-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/XMLHttpRequest/resources/auth1/corsenabled.py. Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.

PASS XMLHttpRequest: send() - "Basic" authenticated CORS requests with user name and password passed to open() (asserts failure) 
